Import: re-implement GPX parsing.
